Fixed #623
popActivity not working as expected
There was a problem with popActivity workflow. In previous version it
never went in first “if condition” because of block in “-
(void)showProgress:(float)progress status:(NSString*)status;” function,
and worked not as in README file described;
My changes helps to make decrement of “activityCount” in case,
described in #623 issue instead of dismissing, and then everything
works fine.

(cherry picked from commit 48800f7)